Initial commit for 6502 emulator
Includes 3 scripts:
* assemble - using dasm (not included here) to assemble a binary file
to be loaded by the program
* run_unit_tests - compiles and executes 6502_unit_tests.c which tests
each of the CPU instructions
* build - compiles and executes 6502_emulator.c, which runs the emulator

The emulator reads in a binary file (currently defined directly in the
code) to memory at a given address and executes the file. When the BRK
instruction is read, the program prints the status of all the registers
to STDOUT and exits. This is used temporarily as a HALT function.