Fix broken Jump instructions
JMP has a bug where, if the indirect address crosses a page, the upper
Byte of the jump address is not read from the next page, but is read
from the first Byte of the current page.

JSR was pushing the wrong value for PC onto the stack. It actually
pushes the address before the next instruction, then RTS increments
after popping to get to the correct address.